more linux-only tests
