+++ cli/tests/host/test.go
