+++ cli/tests/refresh_linux_test.go
@@ -3,8 +3,8 @@ package tests
 import (
 	"testing"
 
+	"github.com/fornellas/resonance/cli/tests/resources"
 	"github.com/fornellas/resonance/resource"
-	"github.com/fornellas/resonance/tests/resources"
 )
 
 func TestRefreshNoPreviousState(t *testing.T) {
