+++ cli/tests/resources/mergeable.go
