+++ cli/tests/restore_linux_test.go
@@ -1,17 +1,18 @@
 package tests
 
 import (
+	"errors"
 	"testing"
 
+	"github.com/fornellas/resonance/cli/tests/resources"
 	"github.com/fornellas/resonance/resource"
-	"github.com/fornellas/resonance/tests/resources"
 )
 
-func TestCheckNoPreviousState(t *testing.T) {
+func TestRestoreNoPreviousState(t *testing.T) {
 	stateRoot, _ := setupDirs(t)
 
 	args := []string{
-		"check",
+		"restore",
 		"--log-level=trace",
 		"--force-color",
 		"--localhost",
@@ -21,15 +22,19 @@ func TestCheckNoPreviousState(t *testing.T) {
 	runCommand(t, Cmd{
 		Args:             args,
 		ExpectedCode:     1,
-		ExpectedInOutput: "No previously saved host state available to check",
+		ExpectedInOutput: "No previously saved host state available to restore from",
 	})
 }
 
-func TestCheckClean(t *testing.T) {
+func TestRestore(t *testing.T) {
 	stateRoot, resourcesRoot := setupDirs(t)
 
 	fooState := resources.IndividualState{
 		Value: "foo",
+	}
+
+	fooStateBroken := resources.IndividualState{
+		Value: "fooBroken",
 	}
 
 	t.Run("apply", func(t *testing.T) {
@@ -79,8 +84,7 @@ func TestCheckClean(t *testing.T) {
 		return
 	}
 
-	t.Run("check is clean", func(t *testing.T) {
-
+	t.Run("restore", func(t *testing.T) {
 		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
 			// Loading saved host state
 			{ValidateName: &resources.IndividualFuncValidateName{
@@ -89,11 +93,20 @@ func TestCheckClean(t *testing.T) {
 			// Reading Host State
 			{GetState: &resources.IndividualFuncGetState{
 				Name:        "foo",
+				ReturnState: fooStateBroken,
+			}},
+			// Executing plan
+			{Configure: &resources.IndividualFuncConfigure{
+				Name:  "foo",
+				State: fooState,
+			}},
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
 				ReturnState: fooState,
 			}},
 		})
 		args := []string{
-			"check",
+			"restore",
 			"--log-level=trace",
 			"--force-color",
 			"--localhost",
@@ -101,16 +114,58 @@ func TestCheckClean(t *testing.T) {
 		}
 		runCommand(t, Cmd{
 			Args:             args,
-			ExpectedInOutput: "State is clean",
+			ExpectedInOutput: "Restore successful",
+		})
+	})
+
+	if t.Failed() {
+		return
+	}
+
+	t.Run("apply", func(t *testing.T) {
+		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
+			// Loading resources
+			{ValidateName: &resources.IndividualFuncValidateName{
+				Name: "foo",
+			}},
+			// Loading saved host state
+			{ValidateName: &resources.IndividualFuncValidateName{
+				Name: "foo",
+			}},
+			// Reading host state
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: fooState,
+			}},
+		})
+		args := []string{
+			"apply",
+			"--log-level=trace",
+			"--force-color",
+			"--localhost",
+			"--state-root", stateRoot,
+			resourcesRoot,
+		}
+		runCommand(t, Cmd{
+			Args:             args,
+			ExpectedInOutput: "Apply successful",
 		})
 	})
 }
 
-func TestCheckDirty(t *testing.T) {
+func TestRestoreFailureWithRollback(t *testing.T) {
 	stateRoot, resourcesRoot := setupDirs(t)
 
 	fooState := resources.IndividualState{
 		Value: "foo",
+	}
+
+	fooStateBroken := resources.IndividualState{
+		Value: "fooBroken",
+	}
+
+	fooStateBad := resources.IndividualState{
+		Value: "fooBad",
 	}
 
 	t.Run("apply", func(t *testing.T) {
@@ -160,8 +215,7 @@ func TestCheckDirty(t *testing.T) {
 		return
 	}
 
-	t.Run("check is dirty", func(t *testing.T) {
-
+	t.Run("restore", func(t *testing.T) {
 		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
 			// Loading saved host state
 			{ValidateName: &resources.IndividualFuncValidateName{
@@ -169,14 +223,33 @@ func TestCheckDirty(t *testing.T) {
 			}},
 			// Reading Host State
 			{GetState: &resources.IndividualFuncGetState{
-				Name: "foo",
-				ReturnState: resources.IndividualState{
-					Value: "fooDirty",
-				},
+				Name:        "foo",
+				ReturnState: fooStateBroken,
+			}},
+			// Executing plan
+			{Configure: &resources.IndividualFuncConfigure{
+				Name:        "foo",
+				State:       fooState,
+				ReturnError: errors.New("fooFailed"),
+			}},
+			// Rollback: Reading host state
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: fooStateBad,
+			}},
+			// Rollback: Executing plan
+			{Configure: &resources.IndividualFuncConfigure{
+				Name:  "foo",
+				State: fooStateBroken,
+			}},
+			// Rollback: Reading host state
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: fooStateBroken,
 			}},
 		})
 		args := []string{
-			"check",
+			"restore",
 			"--log-level=trace",
 			"--force-color",
 			"--localhost",
@@ -185,7 +258,41 @@ func TestCheckDirty(t *testing.T) {
 		runCommand(t, Cmd{
 			Args:             args,
 			ExpectedCode:     1,
-			ExpectedInOutput: "Previous host state is not clean:",
+			ExpectedInOutput: "Failed, rollback to previously saved state successful.",
+		})
+	})
+
+	if t.Failed() {
+		return
+	}
+
+	t.Run("apply", func(t *testing.T) {
+		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
+			// Loading resources
+			{ValidateName: &resources.IndividualFuncValidateName{
+				Name: "foo",
+			}},
+			// Loading saved host state
+			{ValidateName: &resources.IndividualFuncValidateName{
+				Name: "foo",
+			}},
+			// Reading host state
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: fooState,
+			}},
+		})
+		args := []string{
+			"apply",
+			"--log-level=trace",
+			"--force-color",
+			"--localhost",
+			"--state-root", stateRoot,
+			resourcesRoot,
+		}
+		runCommand(t, Cmd{
+			Args:             args,
+			ExpectedInOutput: "Apply successful",
 		})
 	})
 }
