+++ cli/tests/rollback_linux_test.go
@@ -4,15 +4,15 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/fornellas/resonance/cli/tests/resources"
 	"github.com/fornellas/resonance/resource"
-	"github.com/fornellas/resonance/tests/resources"
 )
 
-func TestRestoreNoPreviousState(t *testing.T) {
+func TestRollbackNoPreviousState(t *testing.T) {
 	stateRoot, _ := setupDirs(t)
 
 	args := []string{
-		"restore",
+		"rollback",
 		"--log-level=trace",
 		"--force-color",
 		"--localhost",
@@ -22,15 +22,95 @@ func TestRestoreNoPreviousState(t *testing.T) {
 	runCommand(t, Cmd{
 		Args:             args,
 		ExpectedCode:     1,
-		ExpectedInOutput: "No previously saved host state available to restore from",
+		ExpectedInOutput: "No previously saved host state available to rollback from",
 	})
 }
 
-func TestRestore(t *testing.T) {
+func TestRollbackNoPreviousRollbackState(t *testing.T) {
 	stateRoot, resourcesRoot := setupDirs(t)
 
 	fooState := resources.IndividualState{
 		Value: "foo",
+	}
+
+	t.Run("apply", func(t *testing.T) {
+		setupBundles(t, resourcesRoot, map[string]resource.Resources{
+			"test.yaml": resource.Resources{
+				{
+					TypeName: "Individual[foo]",
+					State:    fooState,
+				},
+			},
+		})
+		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
+			// Loading resources
+			{ValidateName: &resources.IndividualFuncValidateName{
+				Name: "foo",
+			}},
+			// Reading Host State
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: nil,
+			}},
+			// Executing plan
+			{Configure: &resources.IndividualFuncConfigure{
+				Name:  "foo",
+				State: fooState,
+			}},
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: fooState,
+			}},
+		})
+		args := []string{
+			"apply",
+			"--log-level=trace",
+			"--force-color",
+			"--localhost",
+			"--state-root", stateRoot,
+			resourcesRoot,
+		}
+		runCommand(t, Cmd{
+			Args:             args,
+			ExpectedInOutput: "Apply successful",
+		})
+	})
+
+	if t.Failed() {
+		return
+	}
+
+	t.Run("rollback not required", func(t *testing.T) {
+		args := []string{
+			"rollback",
+			"--log-level=trace",
+			"--force-color",
+			"--localhost",
+			"--state-root", stateRoot,
+		}
+		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
+			// Loading resources
+			{ValidateName: &resources.IndividualFuncValidateName{
+				Name: "foo",
+			}},
+		})
+		runCommand(t, Cmd{
+			Args:             args,
+			ExpectedCode:     1,
+			ExpectedInOutput: "No rollback required for saved host state",
+		})
+	})
+}
+
+func TestRollback(t *testing.T) {
+	stateRoot, resourcesRoot := setupDirs(t)
+
+	fooState := resources.IndividualState{
+		Value: "foo",
+	}
+
+	fooStateNew := resources.IndividualState{
+		Value: "fooNew",
 	}
 
 	fooStateBroken := resources.IndividualState{
@@ -84,96 +164,12 @@ func TestRestore(t *testing.T) {
 		return
 	}
 
-	t.Run("restore", func(t *testing.T) {
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading saved host state
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooStateBroken,
-			}},
-			// Executing plan
-			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "foo",
-				State: fooState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-		})
-		args := []string{
-			"restore",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedInOutput: "Restore successful",
-		})
-	})
-
-	if t.Failed() {
-		return
-	}
-
-	t.Run("apply", func(t *testing.T) {
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading resources
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			// Loading saved host state
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			// Reading host state
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-		})
-		args := []string{
-			"apply",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-			resourcesRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedInOutput: "Apply successful",
-		})
-	})
-}
-
-func TestRestoreFailureWithRollback(t *testing.T) {
-	stateRoot, resourcesRoot := setupDirs(t)
-
-	fooState := resources.IndividualState{
-		Value: "foo",
-	}
-
-	fooStateBroken := resources.IndividualState{
-		Value: "fooBroken",
-	}
-
-	fooStateBad := resources.IndividualState{
-		Value: "fooBad",
-	}
-
-	t.Run("apply", func(t *testing.T) {
+	t.Run("apply broken rollback", func(t *testing.T) {
 		setupBundles(t, resourcesRoot, map[string]resource.Resources{
 			"test.yaml": resource.Resources{
 				{
 					TypeName: "Individual[foo]",
-					State:    fooState,
+					State:    fooStateNew,
 				},
 			},
 		})
@@ -182,83 +178,83 @@ func TestRestoreFailureWithRollback(t *testing.T) {
 			{ValidateName: &resources.IndividualFuncValidateName{
 				Name: "foo",
 			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: nil,
-			}},
-			// Executing plan
-			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "foo",
-				State: fooState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-		})
-		args := []string{
-			"apply",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-			resourcesRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedInOutput: "Apply successful",
-		})
-	})
-
-	if t.Failed() {
-		return
-	}
-
-	t.Run("restore", func(t *testing.T) {
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading saved host state
+			// Loading saved state
 			{ValidateName: &resources.IndividualFuncValidateName{
 				Name: "foo",
 			}},
 			// Reading Host State
 			{GetState: &resources.IndividualFuncGetState{
 				Name:        "foo",
-				ReturnState: fooStateBroken,
+				ReturnState: fooState,
 			}},
 			// Executing plan
 			{Configure: &resources.IndividualFuncConfigure{
 				Name:        "foo",
-				State:       fooState,
-				ReturnError: errors.New("fooFailed"),
+				State:       fooStateNew,
+				ReturnError: errors.New("fooFail"),
 			}},
 			// Rollback: Reading host state
 			{GetState: &resources.IndividualFuncGetState{
 				Name:        "foo",
-				ReturnState: fooStateBad,
+				ReturnState: fooStateBroken,
 			}},
 			// Rollback: Executing plan
 			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "foo",
-				State: fooStateBroken,
-			}},
-			// Rollback: Reading host state
-			{GetState: &resources.IndividualFuncGetState{
 				Name:        "foo",
-				ReturnState: fooStateBroken,
+				State:       fooState,
+				ReturnError: errors.New("fooFailAgain"),
 			}},
 		})
 		args := []string{
-			"restore",
+			"apply",
 			"--log-level=trace",
 			"--force-color",
 			"--localhost",
 			"--state-root", stateRoot,
+			resourcesRoot,
 		}
 		runCommand(t, Cmd{
 			Args:             args,
 			ExpectedCode:     1,
-			ExpectedInOutput: "Failed, rollback to previously saved state successful.",
+			ExpectedInOutput: "Rollback failed",
+		})
+	})
+
+	if t.Failed() {
+		return
+	}
+
+	t.Run("rollback", func(t *testing.T) {
+		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
+			// Loading saved host state
+			{ValidateName: &resources.IndividualFuncValidateName{
+				Name: "foo",
+			}},
+			// Reading Host State
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: fooStateBroken,
+			}},
+			// Executing plan
+			{Configure: &resources.IndividualFuncConfigure{
+				Name:  "foo",
+				State: fooState,
+			}},
+			{GetState: &resources.IndividualFuncGetState{
+				Name:        "foo",
+				ReturnState: fooState,
+			}},
+		})
+		args := []string{
+			"rollback",
+			"--log-level=trace",
+			"--force-color",
+			"--localhost",
+			"--state-root", stateRoot,
+		}
+		runCommand(t, Cmd{
+			Args:             args,
+			ExpectedInOutput: "Rollback successful",
 		})
 	})
 
@@ -267,16 +263,24 @@ func TestRestoreFailureWithRollback(t *testing.T) {
 	}
 
 	t.Run("apply", func(t *testing.T) {
+		setupBundles(t, resourcesRoot, map[string]resource.Resources{
+			"test.yaml": resource.Resources{
+				{
+					TypeName: "Individual[foo]",
+					State:    fooState,
+				},
+			},
+		})
 		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
 			// Loading resources
 			{ValidateName: &resources.IndividualFuncValidateName{
 				Name: "foo",
 			}},
-			// Loading saved host state
+			// Loading resources
 			{ValidateName: &resources.IndividualFuncValidateName{
 				Name: "foo",
 			}},
-			// Reading host state
+			// Reading Host State
 			{GetState: &resources.IndividualFuncGetState{
 				Name:        "foo",
 				ReturnState: fooState,
