--- tests/destroy_test.go
@@ -1,332 +1,0 @@
-package tests
-
-import (
-	"errors"
-	"testing"
-
-	"github.com/fornellas/resonance/resource"
-	"github.com/fornellas/resonance/tests/resources"
-)
-
-func TestDestroyNoPreviousState(t *testing.T) {
-	stateRoot, _ := setupDirs(t)
-
-	args := []string{
-		"destroy",
-		"--log-level=trace",
-		"--force-color",
-		"--localhost",
-		"--state-root", stateRoot,
-	}
-
-	runCommand(t, Cmd{
-		Args:             args,
-		ExpectedCode:     1,
-		ExpectedInOutput: "No previously saved host state available to be destroyed",
-	})
-}
-
-func TestDestroy(t *testing.T) {
-	stateRoot, resourcesRoot := setupDirs(t)
-
-	fooState := resources.IndividualState{
-		Value: "foo",
-	}
-
-	t.Run("apply", func(t *testing.T) {
-		setupBundles(t, resourcesRoot, map[string]resource.Resources{
-			"test.yaml": resource.Resources{
-				{
-					TypeName: "Individual[foo]",
-					State:    fooState,
-				},
-			},
-		})
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading resources
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: nil,
-			}},
-			// Executing plan
-			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "foo",
-				State: fooState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-		})
-		args := []string{
-			"apply",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-			resourcesRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedInOutput: "Apply successful",
-		})
-	})
-
-	if t.Failed() {
-		return
-	}
-
-	t.Run("destroy", func(t *testing.T) {
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading saved host state
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-			// Executing plan
-			{Destroy: &resources.IndividualFuncDestroy{
-				Name: "foo",
-			}},
-		})
-		args := []string{
-			"destroy",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedInOutput: "Success",
-		})
-	})
-}
-
-func TestDestroyDirtyState(t *testing.T) {
-	stateRoot, resourcesRoot := setupDirs(t)
-
-	fooState := resources.IndividualState{
-		Value: "foo",
-	}
-
-	barState := resources.IndividualState{
-		Value: "bar",
-	}
-
-	fooNewState := resources.IndividualState{
-		Value: "fooNew",
-	}
-
-	barNewState := resources.IndividualState{
-		Value: "barNew",
-	}
-
-	t.Run("apply", func(t *testing.T) {
-		setupBundles(t, resourcesRoot, map[string]resource.Resources{
-			"test.yaml": resource.Resources{
-				{
-					TypeName: "Individual[foo]",
-					State:    fooState,
-				},
-				{
-					TypeName: "Individual[bar]",
-					State:    barState,
-				},
-			},
-		})
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading resources
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "bar",
-			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: nil,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "bar",
-				ReturnState: nil,
-			}},
-			// Executing plan
-			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "foo",
-				State: fooState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "bar",
-				State: barState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "bar",
-				ReturnState: barState,
-			}},
-		})
-		args := []string{
-			"apply",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-			resourcesRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedInOutput: "Apply successful",
-		})
-	})
-
-	if t.Failed() {
-		return
-	}
-
-	t.Run("destroy with dirty state", func(t *testing.T) {
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading saved host state
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "bar",
-			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooNewState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "bar",
-				ReturnState: barNewState,
-			}},
-		})
-		args := []string{
-			"destroy",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedCode:     1,
-			ExpectedInOutput: "Previous host state is not clean:",
-		})
-	})
-}
-
-func TestDestroyFailureWithSuccessfulRollback(t *testing.T) {
-	stateRoot, resourcesRoot := setupDirs(t)
-
-	fooState := resources.IndividualState{
-		Value: "foo",
-	}
-
-	t.Run("apply", func(t *testing.T) {
-		setupBundles(t, resourcesRoot, map[string]resource.Resources{
-			"test.yaml": resource.Resources{
-				{
-					TypeName: "Individual[foo]",
-					State:    fooState,
-				},
-			},
-		})
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading resources
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: nil,
-			}},
-			// Executing plan
-			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "foo",
-				State: fooState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-		})
-		args := []string{
-			"apply",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-			resourcesRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedInOutput: "Apply successful",
-		})
-	})
-
-	if t.Failed() {
-		return
-	}
-
-	t.Run("destroy with rollback", func(t *testing.T) {
-		resources.SetupIndividualType(t, []resources.IndividualFuncCall{
-			// Loading saved host state
-			{ValidateName: &resources.IndividualFuncValidateName{
-				Name: "foo",
-			}},
-			// Reading Host State
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-			// Executing plan
-			{Destroy: &resources.IndividualFuncDestroy{
-				Name:        "foo",
-				ReturnError: errors.New("fooError"),
-			}},
-			// Rollback: Reading host state
-			{GetState: &resources.IndividualFuncGetState{
-				Name: "foo",
-				ReturnState: resources.IndividualState{
-					Value: "fooError",
-				},
-			}},
-			// Rollback: Applying changes
-			{Configure: &resources.IndividualFuncConfigure{
-				Name:  "foo",
-				State: fooState,
-			}},
-			{GetState: &resources.IndividualFuncGetState{
-				Name:        "foo",
-				ReturnState: fooState,
-			}},
-		})
-		args := []string{
-			"destroy",
-			"--log-level=trace",
-			"--force-color",
-			"--localhost",
-			"--state-root", stateRoot,
-		}
-		runCommand(t, Cmd{
-			Args:             args,
-			ExpectedCode:     1,
-			ExpectedInOutput: "Failed, rollback to previously saved state successful",
-		})
-	})
-}
